chore: moving internal StreamingProcessor class from js-sdk-common to js-server-sdk-common
