+++ packages/shared/common/src/datasource/errors.ts
@@ -35,3 +35,5 @@ export class LDStreamingError extends Error {
     this.recoverable = recoverable;
   }
 }
+
+export type StreamingErrorHandler = (err: LDStreamingError) => void;
