+++ packages/shared/common/src/datasource/index.ts
@@ -1,4 +1,15 @@
 import { DataSourceErrorKind } from './DataSourceErrorKinds';
-import { LDFileDataSourceError, LDPollingError, LDStreamingError } from './errors';
+import {
+  LDFileDataSourceError,
+  LDPollingError,
+  LDStreamingError,
+  StreamingErrorHandler,
+} from './errors';
 
-export { DataSourceErrorKind, LDFileDataSourceError, LDPollingError, LDStreamingError };
+export {
+  DataSourceErrorKind,
+  LDFileDataSourceError,
+  LDPollingError,
+  LDStreamingError,
+  StreamingErrorHandler,
+};
