+++ packages/shared/common/src/internal/index.ts
@@ -2,4 +2,3 @@ export * from './context';
 export * from './diagnostics';
 export * from './evaluation';
 export * from './events';
-export * from './stream';
