--- packages/shared/common/src/internal/stream/index.ts
@@ -1,5 +1,0 @@
-import StreamingProcessor from './StreamingProcessor';
-import type { StreamingErrorHandler } from './types';
-
-export { StreamingProcessor };
-export type { StreamingErrorHandler };
