--- packages/shared/common/src/internal/stream/types.ts
@@ -1,3 +1,0 @@
-import { LDStreamingError } from '../../datasource/errors';
-
-export type StreamingErrorHandler = (err: LDStreamingError) => void;
